Only extract downloaded current tar archive
The current postinst has the bug that all tar archives store in /opt
that begin with ideaIC- are extracted. While this issue will not appear
to most users it may still be a problem to aborted installations which
are only continued after an update.

See also https://gitlab.com/mmk2410/intellij-idea-community/-/issues/5